refactorizando eliminar productos del carrito
